requirements.txt
numpy
aiohttp
Pillow
